Worked on the Game unit tests

+++ internal/model.go
@@ -1,7 +1,7 @@
 package internal
 
 type Model struct {
-	ID        uint32
-	CreatedAt uint32
-	UpdatedAt uint32
+	ID        int64
+	CreatedAt int
+	UpdatedAt int
 }
